+++ .env
@@ -1,0 +1,1 @@
+EXPO_PUBLIC_MAPS_API=AIzaSyC-QXb1OpjIBw8R9TPkZ3Tppr2qiayn2cg
