+++ app/(tabs)/index.tsx
@@ -1,17 +1,5 @@
-import {SafeAreaView, StyleSheet, Text} from 'react-native';
+import HomeScreen from "@/screens/home";
 
 export default function Index() {
-    return (
-        <SafeAreaView style={styles.container}>
-            <Text>Home screen</Text>
-        </SafeAreaView>
-    );
+    return <HomeScreen />;
 }
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        alignItems: 'center',
-        justifyContent: 'center',
-    }
-});
