+++ screens/home/index.tsx
@@ -1,0 +1,10 @@
+import {SafeAreaView, Text} from 'react-native';
+import styles from './styles'
+
+export default function HomeScreen() {
+    return (
+        <SafeAreaView style={styles.container}>
+            <Text>Home screen</Text>
+        </SafeAreaView>
+    );
+}
