+++ screens/home/styles.tsx
@@ -1,0 +1,11 @@
+import {StyleSheet} from "react-native";
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        alignItems: 'center',
+        justifyContent: 'center',
+    }
+});
+
+export default styles
